+++ quantlib/termstructures/yields/rate_helpers.pyx
@@ -120,7 +120,7 @@ cdef class SwapRateHelper(RelativeDateRateHelper):
         self._thisptr = ptr
 
     @classmethod
-    def from_tenor(cls, float rate, Period tenor,
+    def from_tenor(cls, double rate, Period tenor,
         Calendar calendar, Frequency fixedFrequency,
         BusinessDayConvention fixedConvention, DayCounter fixedDayCount,
         IborIndex iborIndex, Quote spread=None, Period fwdStart=None):
@@ -161,7 +161,7 @@ cdef class SwapRateHelper(RelativeDateRateHelper):
         return instance
 
     @classmethod
-    def from_index(cls, float rate, SwapIndex index):
+    def from_index(cls, double rate, SwapIndex index):
 
         cdef Handle[_qt.Quote] spread_handle = Handle[_qt.Quote](new _qt.SimpleQuote(0))
         cdef Period p = Period(2, Days)
